Moved DesignTime projects - they mysteriously ended up in the Solutions folder
Position, Target on Camera hooked up for Xaml

Implemented TypeConverter for Vector - one can now typically do :
Position="0,0,0" on a Node (Mesh, Camera, Light...)

Added OmniLight as a Node.

Also done a default InternalGame for those scenarios were one don't need a game
The internal infrastructure is tied into the concept of having a Game for
rendering purposes - until we have "passive" rendering, this is the easiest
way to do it.